move mutation search to own file
